+++ sklearn/__init__.py
@@ -67,7 +67,8 @@ else:
                'feature_selection', 'semi_supervised',
                'gaussian_process', 'grid_search', 'hmm', 'lda', 'linear_model',
                'metrics', 'mixture', 'naive_bayes', 'neighbors', 'pipeline',
-               'preprocessing', 'qda', 'svm', 'test', 'clone', 'pls']
+               'preprocessing', 'qda', 'svm', 'test', 'clone', 'pls',
+               'isotonic']
 
 
 def setup_module(module):