+++ sklearn/tests/test_naive_bayes.py
@@ -92,15 +92,11 @@ def test_discretenb_pickle():
         clf = cls().fit(X2, y2)
         y_pred = clf.predict(X2)
 
-        store = StringIO()
+        store = BytesIO()
         pickle.dump(clf, store)
-        clf = pickle.load(StringIO(store.getvalue()))
+        clf = pickle.load(BytesIO(store.getvalue()))
 
         assert_array_equal(y_pred, clf.predict(X2))
-
-    store = BytesIO()
-    pickle.dump(clf, store)
-    clf = pickle.load(BytesIO(store.getvalue()))
 
 
 def test_input_check():