+++ doc/whats_new.rst
@@ -7,11 +7,12 @@
 
 Changelog
 ---------
+
    - Partial dependence plots for :mod:`ensemble.gradient_boosting` by
      `Peter Prettenhofer`_.
 
-   - New estimators :class: `linear_model.PassiveAggressiveClassifier` and
-     `linear_model.PassiverAggressiveRegressor` by `Rob Zinkov` and
+   - New estimators :class:`linear_model.PassiveAggressiveClassifier` and
+     :class:`linear_model.PassiverAggressiveRegressor` by `Rob Zinkov` and
      `Mathieu Blondel`_.
 
    - The table of contents has now been made expandible (on the
@@ -44,7 +45,7 @@ Changelog
      of several transformers by `Andreas Müller`_.
 
    - Faster and more robust :func:`metrics.confusion_matrix` and
-     :ref:`clustering_evaluation`_ by Wei Li.
+     :ref:`clustering_evaluation` by Wei Li.
 
    - New estimator :class:`decomposition.FactorAnalysis` by
      `Christian Osendorfer`_ and `Alexandre Gramfort`_
@@ -57,6 +58,11 @@ Changelog
    - LARS algorithm made more numerically stable with heuristics to drop
      regressors too correlated as well as to stop the path when
      numerical noise becomes predominant, by `Gael Varoquaux`_.
+
+   - New estimator :class:`preprocessing.OneHotEncoder` to compute
+     binary encodings of categorical features by `Andreas Müller`_.
+
+   - Faster implementation of :func:`metrics.precision_recall_curve` by Conrad Lee.
 
 
 API changes summary